+++ packages/checks/markdown-front-matter/vitest.config.mjs
@@ -4,5 +4,6 @@ export default defineConfig({
   esbuild: { target: 'ES2022' },
   test: {
     globals: true,
+    include: ['src/**/*.test.ts'],
   },
 });