+++ packages/checks/prettier/vitest.config.mjs
@@ -8,7 +8,7 @@ export default defineConfig({
   esbuild: { target: 'ES2022' },
   resolve: {
     alias: {
-      '@mayjournal/fitness': join(packageRoot, '../../runner/dist/src/types/index.types.js'),
+      '@mayjournal/fitness': join(packageRoot, '../../runner/dist/types/index.types.js'),
     },
   },
   test: {
@@ -25,5 +25,6 @@ export default defineConfig({
       },
     },
     globals: true,
+    include: ['src/**/*.test.ts'],
   },
 });